Feito funcoes para tela de categoria, ja funcionando porem nao foi realizado os testes e nem criado JUnit

+++ backend/demo/src/main/java/com/back/demo/repository/CategoriaDTORepository.java
@@ -28,7 +28,7 @@ public class CategoriaDTORepository {
         return q.getSingleResult();
     }
 
-    public Long getCountProductByCategoria(Integer categoriaId, String status, String search){
+    public Long getCountProductByCategoria(Long categoriaId, String status, String search){
         String query = "SELECT COUNT(i) " + 
                        "FROM  Item i " +
                        "JOIN Categoria c ON i.id = c.id ";
